Engadget: Pikmin 4 is getting a free update with hard mode, Decor Pikmin and a camera to snap field photos
But there's no Switch 2 upgrade (yet).

Cheyenne MacDonald

Weekend Editor

Wed, October 22, 2025 at 4:52 PM UTC

Nintendo

Nintendo has announced a free update for Pikmin 4 that could remedy one of fans' biggest complaints about the game. (Yes, free. From Nintendo. We're shocked too, but we'll take it.) The update, which arrives in November, adds Fierce mode to make the game more challenging for those who want it. We don't know exactly what that will look like yet, but a higher difficulty level is a welcome addition to a game that is, for many players including myself, absolutely delightful but much too easy. It may not reach Pikmin 2 heights still, but maybe we'll see enemies respawn?

On the opposite end, for players who want to breeze through an even chiller experience than Pikmin 4 already offered, the update also brings Relaxed mode, where enemies won't bother you unless you attack them first. This mode could go hand in hand with another new feature, the Field Camera, which puts a Pokémon-Snap-style spin on things. With this, you'll be able to take photos of your Pikmin and squad in the wild, and "set the scene with various angles, hide characters, and decorate your pics with filters, frames and stamps." It might be nice not to have enemies pestering you while you're just trying to have a photoshoot.

The update also brings a Pikmin Bloom staple over to the mainline game: Decor Pikmin. You'll soon be able to find Pikmin that are decked out in little costumes, and can add these characters to your squad. The announcement hints that snapping a picture of them may be the way to do this. And, Pikmin Bloom players will be able to send these special Decor Pikmin to the app to join their squad over there too. There's more news for Pikmin Bloom players as well. The mobile game is finally getting Ice Pikmin on November 1. These will sprout from new blue seedlings that will be found, to start, at Restaurant and Roadside locations.

An Ice Pikmin with a scoop of vanilla ice cream on its head and a pink spoon sticking out sits on a rocky ledge

(Nintendo)

At the beginning of October, Nintendo got us all worked up about the possibility of new Pikmin content when it dropped an animated short that at first seemed to be Pikmin related, and was later confirmed to be. The company eventually put the hype to rest, explaining that the videos were indeed just the first short films from Nintendo Pictures Co., which will "explore new creative possibilities through video content" — not a promotion for a new game, as fans hoped. But with a new reason to dive back into Pikmin 4 now on the horizon, maybe all the excitement wasn't for nothing.

Advertisement

About our ads