Techcrunch:
Rivians founder and CEO RJ Scaringe is taking on a new role at his company as part of a restructuring ahead of next years launch of the R2 SUV, which involves a layoff of more than 600 employees.

Scaringe told employees in an email on Thursday that he will be acting as Rivians chief marketing officer on an interim basis as the company searches for someone to fill that position for the first time. The head of Rivians &#8220;marketing experiences&#8221; team and the lead of the companys creative studio will now report directly to Scaringe, according to a copy of the email obtained by TechCrunch.