Techcrunch:
Rivian has agreed to pay $250 million to settle a class action shareholder lawsuit filed after the company suddenly hiked prices on its R1 pickup truck and SUV in 2022.

The lawsuit alleged Rivian had included misleading statements and figures in regulatory filings in the run-up to its 2021 IPO about the costs required to build the R1 EVs. Despite agreeing to the payment, Rivian said in a press release that it &#8220;denies the allegations in the suit and maintains that this agreement to settle is not an admission of fault or wrongdoing.&#8221;