Engadget: Deals
Best iPad deals: Get over $300 off the iPad Air M3 with cellular

Other Apple deals include discounts on a four-pack of AirTags.

Valentina Palladino, Amy Skorheim, Jeff Dunn

UpdatedFri, October 24, 2025 at 4:39 PM UTC

Nathan Ingraham for Engadget

The just-released iPad Pro with the M5 chip tops our list of the best tablets and the standard iPad is our pick for the best budget slate. While the former is expectedly not on sale yet, we are seeing a modest discount for the cheaper iPad. The lovely iPad Air (13-inch, with cellular) is down to a record low as well. Of course, you won't find deals on Apple's own website, but we keep an eye on Amazon, Target, Walmart and other retailers to find the best iPad deals out there and round them up each Friday. This week, the discounts aren't as good as they were for Prime Day earlier this month — chances are, we won't see a huge influx of Apple deals until Black Friday sales start up. Until then, here are the top deals on iPads and all the other Apple gear we could find.

Best iPad deals

Apple

iPad Air (13-inch, M3, 512GB, Wi-Fi + cellular, Starlight)

$941$1,249Save $308

The most recent iPad Air in the 13-inch size with cellular connectivity and 512GB of storage is currently cheaper than the Wi-Fi only configuration. The discount only applies to the Starlight colorway, but if that color appeals to you, you'll save $350 over the MSRP.

Engadget's Nate Ingraham gave the 13-inch iPad Air a score of 89 when it was released in March. It has a bigger and slightly brighter display than its 11-inch counterpart but is otherwise the same. If you plan to keep your iPad hooked up to a keyboard, the extra screen space is lovely for multitasking or just taking in movies.

$941 at Amazon

Explore More Buying Options

$1,250 at Verizon

Apple iPad (A16, 256GB) for $399 ($50 off): The latest entry-level iPad comes with a faster A16 chip, 2GB more RAM and more base storage. It earned a score of 84 in our review — if you only need a tablet for roaming the internet, watching shows and doing some lighter productivity tasks, it should do the job. With the recent iPadOS 26 update, it also has most of the same multitasking features available on the more expensive models. It does lack Apple Intelligence, but to be candid, that isn't a big loss right now. This deal isn't an all-time low for the model with 256GB of storage but it takes $50 off Apple's list price. Also at Best Buy.

Apple iPad Air (11-inch, M3, 1TB) for $949 ($150 off MSRP): The most recent iPad Air is a relatively minor update, as the only major addition is a more powerful M3 chip. However, we still recommend the Air over the base model in our iPad buying guide: Its display is laminated, more color-rich and better at fending off glare (though it's still 60Hz); its speakers are more robust; it works with Apples best accessories and its performance should hold up better in the years ahead. This deal is only for the maxed-out model with 1TB of storage, but it ties the lowest price we've seen all the same.

Best Apple deals

Apple

Apple AirTag 4 Pack

$65$99Save $34

The best Bluetooth tracker for iPhones is still the AirTag. We're waiting for a refreshed version to come out, but after announcing the iPhone 17 in September, followed by devices with the new M5 chip in October, Apple hasn't revealed new tags. These are still eerily accurate in finding lost items using the vast network of anonymous iPhones and the Find My app — just keep in mind you'll need an accessory to attach the discs, as these are one of the few trackers out there that doesn't have an attachment point. This price matches the lowest we've seen on the tags so far.

$65 at Amazon

Explore More Buying Options

$100 at Target$100 at Verizon

Apple Pencil Pro for $99 ($30 off): The top-end option in Apples confusing stylus lineup, the Pencil Pro supports pressure sensitivity, wireless charging, tilt detection, haptic feedback and Apples double tap and squeeze gestures, among other perks. Its a lovely tool for more intricate sketching and note-taking, but the catch is that its only compatible with the M4 iPad Pro, M2 and M3 iPad Air and most recent iPad mini. We've seen this deal fairly often over the course of the year, but it's a fine discount compared to buying from Apple directly. Also at Walmart.

Advertisement

Advertisement

Apple MacBook Air (13-inch, M4, 512GB) for $999 ($200 off): Apple's latest MacBook Air is the top pick in our guide to the best laptops, and it earned a score of 92 in our review. It's not a major overhaul, but the design is still exceptionally thin, light and well-built, with long battery life and a top-notch keyboard and trackpad. Now it's a bit faster. (Though we'd still love more ports and a refresh rate higher than 60Hz.) This discount ties the all-time low for the model with 16GB of RAM and a 512GB SSD.

Apple Watch Series 11 (GPS, 42mm) for $389 ($10 off): The latest flagship Apple Watch only hit store shelves last month, but Amazon is already selling it for $10 off. It doesn't show up as a percentage off, but you'll see some models listed at $389 instead of Apple's $399 MSRP. If you're new to Apple's wearables or are ready to upgrade from a Series 9 or older, this is a good model to grab. If you're coming from a Series 10, however, there's not much need to upgrade as the only major change from last year's model is a slightly larger battery and a tougher screen.

Apple Watch SE 3 (GPS, 40mm) for $240 ($9 off): There's a similar stealth discount for the newest budget model, the Apple Watch SE 3, at Amazon. It normally goes for $249 — again, not a big discount, but better than nothing if you're looking to get onboard early. Apple gave this model some badly needed updates compared to its predecessor, including an always-on display, faster charging, better sensors and the same processor that you'll find in the new Apple Watch Series 11.

Read more Apple coverage:

The best AirPods

The best Apple Watches

The best MacBooks

The best iPhones

The best iPads

Follow @EngadgetDeals on X for the latest tech deals and buying advice.

Advertisement

About our ads