Techcrunch:
OpenAI is working on a new tool that would generate music based on text and audio prompts, according to a report in The Information.

Such a tool could be used to add music to existing videos, or to add guitar accompaniment to an existing vocal track, sources said. Its not clear when OpenAI plans to launch such a tool or whether it would be available as a standalone product (as opposed to integrating with OpenAIs ChatGPT and video app Sora).