Techcrunch:
Every few weekends, 21-year-old student Lavanya Jain opens the BlaBlaCar app to find a lift from Noida on the outskirts of New Delhi to his home in Kandhla, a small town in northern Indias Uttar Pradesh state. The 120-kilometer journey costs him about ₹500, the equivalent of about $6. Thats a fraction of the ₹1,500–₹2,000, or $17–$23, he would pay for a private cab.

&#8220;If youre looking for a fast, efficient, affordable, and comfortable way to travel — and you like to chat — you should basically check out BlaBlaCar,&#8221; Jain told TechCrunch, adding that he has used the app some 40 to 50 times over the past two years.