Techcrunch:
New AI-powered web browsers such as OpenAIs ChatGPT Atlas and Perplexitys Comet are trying to unseat Google Chrome as the front door to the internet for billions of users. A key selling point of these products are their web browsing AI agents, which promise to complete tasks on a users behalf by clicking around on websites and filling out forms.

But consumers may not be aware of the major risks to user privacy that come along with agentic browsing, a problem that the entire tech industry is trying to grapple with.