Techcrunch:
Two years ago, former SpaceX engineers used rocket tech to develop a power plant capable of removing carbon dioxide from the atmosphere, with a small asterisk. To draw down CO2, it burned plant waste, making it a sort of &#8220;vegetarian rocket engine&#8221; for the grid.

This week, Arbor Energy said it had raised a $55 million Series A led by Lowercarbon Capital and Voyager Ventures on the heels of a partial pivot. Its power plant, rather than sticking to a strict vegetarian diet, is going to be more of an omnivore, capable of burning natural gas in addition to biomass.