Engadget: Here's our first look at the Paranormal Activity game from the maker of The Mortuary Assistant
The found footage horror game is called Paranormal Activity: Threshold.

Cheyenne MacDonald

Weekend Editor

Sat, October 25, 2025 at 9:06 PM UTC

A teaser shared at the end of the Indie Horror Showcase this week gives us a better idea of what the upcoming found footage Paranormal Activity game will be like. In the short trailer for Paranormal Activity: Threshold, we're introduced to an unfortunate couple who has purchased a home that not only needs some serious work, but is also very much haunted. Cue creepy faces appearing in doorways and around corners.

While yet another installation in the Paranormal Activity franchise may not be what everyone is clamoring for, the fact that Threshold is being developed by solo dev Brian Clarke (DarkStone Digital), the creator of The Mortuary Assistant, is pretty promising. Per the game's Steam page, you'll be able to "Play in multiple timelines; hunt and communicate with entities; perform demonic rituals to alter your fate;" and "discover multiple endings and expand the story." Each playthrough will be unique, thanks to the "Paranormal Engine," it adds.

As of now, there's no release date, but you can wishlist it on Steam. It was previously said the game would be released in 2026 on multiple platforms, but the Steam page only says it's coming soon.

Advertisement

About our ads