Techcrunch:
Apple Maps users could start seeing ads in the app as soon as next year, according to a new report from Bloombergs Mark Gurman.

Similar to Google Maps and other mapping apps, Apples plan is to allow restaurants and other businesses with brick-and-mortar locations to pay to promote themselves in search results, Gurman says. While Apple already runs ads in the App Store, this could be part of a larger strategy to introduce more advertising in iOS.