Techcrunch:
The United States and China are ready to move forward on a TikTok deal, according to Treasury Secretary Scott Bessent.

Bessent had already said last month that the two countries reached a “framework” on a deal during discussions in Madrid, and President Donald Trump subsequently signed an executive order to facilitate the transaction.