Engadget: OpenAI is reportedly working on an AI music-generation tool
The slop expands.

Cheyenne MacDonald

Weekend Editor

Sun, October 26, 2025 at 8:42 PM UTC

OpenAI

According to a report from The Information, OpenAI is interested in developing a tool that could generate music from text and audio prompts, and has been working with students from The Juilliard School to help with creating training data. The company envisions something that could be used to generate guitar accompaniment for a vocal track, for example, or to add music to videos, sources with knowledge of the project told the publication.

There's no word on how far into these efforts OpenAI is. According to one source that spoke to The Information, OpenAI enlisted the Juilliard students to annotate music scores, which could be used to train the AI. It wouldn't be the first time the company has explored music-generating AI, and it's something we're seeing more and more of elsewhere with other startups, like Suno and ElevenLabs, launching their own versions. AI spam is already cluttering up streaming platforms and, lest we forget, there was the whole The Velvet Sundown debacle. It's only the beginning.

Advertisement

About our ads