Techcrunch:
Plastic recycling has fallen short. Only about 9% of all plastic is recycled globally, which sounds pretty bad until you compare it with textiles. Only 0.5% of those are recycled.

One of the biggest challenges is that textiles are seldom one material. Buttons and zippers complicate matters, but spandex is even worse. Novel synthetic blends have made for clothing thats a dream to wear but a nightmare to recycle.