Techcrunch:
Robots can be programmed to do a variety of tasks, like packing boxes and even performing surgery. But each individual movement or task requires its own specific training process, which makes it hard for robots to adapt in real-world scenarios.

Mbodi wants to make training robots easier and quicker with the help of AI agents. The company will be showcasing this tech as one of the Top 20 Startup Battlefield finalists at TechCrunch Disrupt 2025.