Techcrunch:
Kevin Damoa came face-to-face with the challenges and dangers of moving freight from road to rail as a 17-year-old U.S. Army enlistee tasked with loading tanks and Bradley fighting vehicles onto the railroad. It was — as the mechanical engineer and founder of Glīd Technologies puts it — the beginning of his love story with logistics.

Its a love story that persisted through a 13-year stint with the U.S. Air Force National Guard as a firefighter to his roles in the private sector at SpaceX, Northrup Grumman, Romeo Power Tech, and Xos Trucks — to name a few.