Techcrunch:
Waymos co-CEO, Tekedra Mawakana, had a clear message during her interview on the TechCrunch Disrupt 2025 stage Monday: &#8220;It is imperative that we scale.&#8221;

Mawakana was speaking in the context of how Waymo balances fundraising (and burning through that money) with eventually achieving profitability. But she was also clear in the interview that she believes Waymo can increase road safety by reaching that scale.