Engadget: Deals
Apple's MacBook Air M4 is back on sale for $799

This is a discount of 20 percent.

Lawrence Bonk

Contributing Reporter

Mon, October 27, 2025 at 6:38 PM UTC

Devindra Hardawar for Engadget

The Apple MacBook Air M4 laptop is back on sale for just $799, which is one heck of a deal. This sale is for the model with 16GB of RAM and 256GB of internal storage. It's available in multiple colorways, but the silver one inexplicably costs $899.

We ranked this as our favorite Apple laptop in our list of the best MacBook computers. Heck, it's even our very favorite laptop. Full stop. The performance is exceptionally snappy, thanks to the M4 chip. We appreciated the upgraded battery life, which now lasts for around 18 hours per charge. That's well beyond a full day of work.

Apple

Apple MacBook Air 13-inch Laptop with M4 chip (16GB/256GB)

$799$999Save $200

This is 20 percent off.

$799 at Amazon

Explore More Buying Options

$984 at Best Buy$894 at P.C. Richard & Son

The design is lightweight, but sturdy. This has become a hallmark for modern MacBook Air computers. The screen is both gorgeous and roomy, even though it's technically just a 13-inch panel. There's support for the P3 wide color gamut and it can reach up to 500 nits of brightness.

Advertisement

Advertisement

This is a near-perfect laptop, but there are a couple of nitpicks. There's no USB-C port on the right side, limiting how users can arrange accessories on a desk. Also, the screen is capped with a 60Hz refresh rate. Another potential complication is the looming specter of the M5 chip. The company has already released the MacBook Pro M5, so a new MacBook Air is likely coming in the nearish future.

Check out our coverage of the best Apple deals for more discounts, and follow @EngadgetDeals on X for the latest tech deals and buying advice.

Advertisement

About our ads