Engadget: AI
Googles AI health coach will soon be available to some Fitbit Premium users

It starts with a five or ten minute conversation with the bot to assess fitness goals.

Lawrence Bonk

Contributing Reporter

Mon, October 27, 2025 at 5:00 PM UTC

Google

Googles long-awaited AI health coach is nearly upon us, as a preview version is launching tomorrow for some Fitbit Premium users in the US. This will only be for Android devices at first, but the company promises an iOS version is in the works.

This is a Public Preview version of the software, so think of it like a beta release. Google says it'll incorporate user feedback to "add, change or improve features and capabilities." The company warns users that this is a "new experience, so initially, there will be some gaps."

For the uninitiated, Google's AI health coach is exactly what it sounds like. This is an AI chatbot intended to help users reach fitness and health goals. The company boasts that the tech is "secure, personalized and grounded in science." Everything starts with a five to ten minute conversation with the coach to assess health and fitness goals.

Advertisement

Advertisement

Advertisement

The coach can be a sounding board for personal health, fitness and sleep goals, but also acts as a personal trainer. Google says it can be used to review and adjust fitness plans, check progress, get advice on trends and create workouts. To that last point, the company says the chatbot can create workouts based on pre-existing constraints. For instance, users can ask the bot to make a workout that can be done in a cramped hotel room.

The coach can also be used to brainstorm questions to ask a doctor and to track and analyze a number of sleep metrics. The bot provides a "detailed sleep analysis" and can allegedly understand patterns and trends that can impact sleep. All of this data can be accessed via the app.

Being as this is a preview build, it won't roll out to everyone tomorrow. Eligible Fitbit Premium users will receive notification that the software is ready to use. It works with any Pixel Watch or Fitbit device.

Google

The entire Fitbit app is being redesigned to focus more on AI and this is a large piece of the puzzle. Google promises integration with its health coach across every aspect of the app.

Advertisement

About our ads