Engadget: YouTube TV is giving subscribers a $20 credit as consolation for the Disney blackout
Instructions to redeem the credit should arrive via email in 'the next few days.'

Cheyenne MacDonald

Weekend Editor

Sun, November 9, 2025 at 11:10 PM UTC

NurPhoto via Getty Images

YouTube TV has notified subscribers that a $20 credit is heading their way in light of its ongoing standoff with Disney, which has resulted in numerous channels going dark on the platform, including ESPN, ABC News and Disney Channel. In an email to subscribers, the YouTube TV Team said subscribers should expect to see an email in the coming days about the credit, which will be applied to their next bill after it's redeemed.

"We know its been disappointing to lose Disney content, and we want you to know we deeply appreciate your patience," the email said. YouTube TV and Disney had a deadline of October 30 to reach a new agreement about Disney's content moving forward, but that date passed with no deal. Negotiations are ongoing, according to the company. "We've been working in good faith to negotiate a deal with Disney that pays them fairly for their content and returns their programming to YouTube TV," the team said in the email. If that $20 still isn't enough to make up for the weeks of lost content, you can always pause your subscription in the meantime.

Advertisement

About our ads